Refactored code for the SwiftUI Transactions TableViewCell
- Polished transaction details
- Removed the old UITableViewCells (UIKit)
- SwiftUI Used whenever available
- Added supporting SwiftUI Views
