+++ loafwallet/TransactionCellView.swift
@@ -7,8 +7,12 @@
 //
 
 import SwiftUI
+
  
 struct TransactionCellView: View {
+    
+    private let imageLength: CGFloat = 15.0
+
     
     //MARK: - Combine Variables
     @ObservedObject
@@ -23,46 +27,57 @@ struct TransactionCellView: View {
             VStack(alignment: .leading, spacing: 5.0) {
             
             //Send and Date Labels
-           HStack(alignment: .top, spacing: 1.0) {
+            HStack(alignment: .bottom, spacing: 1.0) {
                  
                 Text(viewModel.amountText)
-                    .font(Font(UIFont.barlowBold(size: 12.0)))
+                    .font(Font(UIFont.barlowSemiBold(size: 14.0)))
                     .foregroundColor(.black)
-                
-               Spacer()
-                
+     
+                Spacer()
+
                 Image(systemName: viewModel.directionImageText)
                     .resizable()
                     .aspectRatio(contentMode: .fit)
-                    .frame(width: 15.0,
-                           height: 15.0)
+                    .frame(width: imageLength,
+                           height: imageLength)
                     .foregroundColor(viewModel.directionArrowColor)
-
+                    .padding(.trailing, 1.0)
+     
                 Text(viewModel.timedateText)
-                    .font(Font(UIFont.barlowRegular(size: 12.0)))
+                    .font(Font(UIFont.barlowRegular(size: 13.0)))
                     .foregroundColor(Color.black)
                     .frame(width: 50.0, alignment: .trailing)
             }
             .padding([.leading,.trailing], 10.0)
-            .padding(.top,5.0)
 
-                Spacer(minLength: 2.0)
+            //Info and Direction arrow
+                HStack(alignment: .center, spacing: 1.0) {
+
+                Text(viewModel.transaction.status)
+                    .font(Font(UIFont.barlowRegular(size: 13.0)))
+                    .foregroundColor(.black)
                 
-            //Address and Status
-            HStack(spacing: 1.0) {
+                Spacer()
+                    
+                Image(systemName: "info.circle")
+                        .resizable()
+                        .aspectRatio(contentMode: .fit)
+                        .frame(width: imageLength,
+                               height: imageLength)
+            }
+            .padding([.leading,.trailing], 10.0)
+
+            //Address
+           HStack(alignment: .top, spacing: 1.0) {
+
                 Text(viewModel.addressText)
                     .truncationMode(.middle)
-                    .font(Font(UIFont.barlowRegular(size: 12.0)))
+                    .font(Font(UIFont.barlowRegular(size: 14.0)))
                     .foregroundColor(.black)
-                    .frame(width: 140.0)
+                    .frame(width: 160.0)
                     .lineLimit(1)
 
                 Spacer()
-                
-                Text(viewModel.transaction.status)
-                    .font(Font(UIFont.barlowRegular(size: 12.0)))
-                    .foregroundColor(.black)
-                 
             }
             .padding([.leading,.trailing], 10.0)
 
@@ -70,7 +85,7 @@ struct TransactionCellView: View {
                 .frame(height: 1.0)
                 .background(Color(UIColor.lightGray))
                 .padding([.leading,.trailing], 10.0)
-                .padding(.bottom,5.0)
+                .padding(.bottom,3.0)
         }
         
     }
