+++ loafwallet/TransactionCellViewModel.swift
@@ -9,8 +9,9 @@
 import Foundation
 import SwiftUI
 
+private let timestampRefreshRate: TimeInterval = 10.0
+
 class TransactionCellViewModel: ObservableObject {
-    
     
     //MARK: - Public Variables
     var transaction: Transaction
@@ -23,11 +24,13 @@ class TransactionCellViewModel: ObservableObject {
     
     var isSyncing: Bool
      
-    var  amountText: String = ""
+    var amountText: String = ""
     
-    var  directionText: String = ""
+    var feeText: String = ""
+  
+    var directionText: String = ""
     
-    var  directionImageText: String = ""
+    var directionImageText: String = ""
     
     var directionArrowColor: Color = .clear
     
@@ -35,7 +38,13 @@ class TransactionCellViewModel: ObservableObject {
     
     var timedateText: String = ""
     
-
+    var memoString: String = ""
+    
+    var qrImage = UIImage()
+    
+    //MARK: - Private Variables
+    private var timer: Timer? = nil
+ 
     init(transaction: Transaction,
          isLtcSwapped: Bool,
          rate: Rate,
@@ -54,6 +63,10 @@ class TransactionCellViewModel: ObservableObject {
     private func loadVariables() {
         
         amountText = transaction.descriptionString(isLtcSwapped: isLtcSwapped, rate: rate, maxDigits: maxDigits).string
+         
+        feeText = transaction.amountDetails(isLtcSwapped: isLtcSwapped, rate: rate, rates: [rate], maxDigits: maxDigits)
+        
+        addressText = String(format: transaction.direction.addressTextFormat, transaction.toAddress ?? "")
         
         if self.transaction.direction == .sent {
             directionImageText = "arrowtriangle.up.circle.fill"
@@ -63,236 +76,26 @@ class TransactionCellViewModel: ObservableObject {
             directionArrowColor = Color(UIColor.litecoinOrange) 
         }
         
-      let timestampInfo = transaction.timeSince
-      timedateText = timestampInfo.0
-      
+        let timestampInfo = transaction.timeSince
+        timedateText = timestampInfo.0
         if timestampInfo.1 {
-                   
-            
-//                timer = Timer.scheduledTimer(timeInterval: timestampRefreshRate, target: TransactionCellWrapper(target: self), selector: NSSelectorFromString("timerDidFire"), userInfo: nil, repeats: true)
-//               } else {
-//                   timer?.invalidate()
-//               }
+            timer = Timer.scheduledTimer(timeInterval: timestampRefreshRate, target: self, selector: NSSelectorFromString("timerDidFire"), userInfo: nil, repeats: true)
+        } else {
+            timer?.invalidate()
+        }
+          
+        if let address = transaction.toAddress,
+           let data = address.data(using: .utf8),
+           let image = UIImage
+                .qrCode(data: data,
+                        color: .black)?
+                .resize(CGSize(width: kQRImageSide,
+                               height: kQRImageSide)) {
+                            qrImage = image
+        }
         
-//               timedateLabel.isHidden = !transaction.isValid
+        if let memo = self.transaction.comment {
+            memoString = memo
+        }
     }
-        
-        addressText = String(format: transaction.direction.addressTextFormat, transaction.toAddress ?? "")
-    }
- 
 }
-
-
-//func setTransaction(_ transaction: Transaction, isLtcSwapped: Bool, rate: Rate, maxDigits: Int, isSyncing: Bool) {
-//
-//    self.transaction = transaction
-//
-//    amountLabel.attributedText = transaction.descriptionString(isLtcSwapped: isLtcSwapped, rate: rate, maxDigits: maxDigits)
-//    addressLabel.text = String(format: transaction.direction.addressTextFormat, transaction.toAddress ?? "")
-//    statusLabel.text = transaction.status
-//    commentTextLabel.text = transaction.comment
-//    blockLabel.text = transaction.blockHeight
-//    txidStringLabel.text = transaction.hash
-//    availabilityLabel.text = transaction.shouldDisplayAvailableToSpend ? S.Transaction.available : ""
-//    arrowImageView.image = UIImage(named: "black-circle-arrow-right")?.withRenderingMode(.alwaysTemplate)
-//    startingBalanceLabel.text = transaction.amountDetailsStartingBalanceString(isLtcSwapped: isLtcSwapped, rate: rate, rates: [rate], maxDigits: maxDigits)
-//    endingBalanceLabel.text = transaction.amountDetailsEndingBalanceString(isLtcSwapped: isLtcSwapped, rate: rate, rates: [rate], maxDigits: maxDigits)
-//    dropArrowImageView.image = UIImage(named: "modeDropArrow")
-//
-//    if #available(iOS 11.0, *) {
-//
-//        guard let textColor = UIColor(named: "labelTextColor") else {
-//            NSLog("ERROR: Custom color not found")
-//            return
-//        }
-//
-//        guard let backgroundColor = UIColor(named: "inverseBackgroundViewColor") else {
-//            NSLog("ERROR: Custom color not found")
-//            return
-//        }
-//
-//        amountLabel.textColor = textColor
-//        addressLabel.textColor = textColor
-//        commentTextLabel.textColor = textColor
-//        statusLabel.textColor = textColor
-//        timedateLabel.textColor = textColor
-//
-//        staticCommentLabel.textColor = textColor
-//        staticTxIDLabel.textColor = textColor
-//        staticAmountDetailLabel.textColor = textColor
-//        staticBlockLabel.textColor = textColor
-//
-//        qrBackgroundView.backgroundColor = backgroundColor
-//
-//    } else {
-//        commentTextLabel.textColor = .darkText
-//        statusLabel.textColor = .darkText
-//        timedateLabel.textColor = .darkText
-//
-//        staticCommentLabel.textColor = .darkText
-//        staticTxIDLabel.textColor = .darkText
-//        staticAmountDetailLabel.textColor = .darkText
-//        staticBlockLabel.textColor = .darkText
-//
-//        qrBackgroundView.backgroundColor = .white
-//    }
-//
-//    if transaction.status == S.Transaction.complete {
-//        statusLabel.isHidden = false
-//    } else {
-//        statusLabel.isHidden = isSyncing
-//    }
-//
-//    let timestampInfo = transaction.timeSince
-//    timedateLabel.text = timestampInfo.0
-//    if timestampInfo.1 {
-//        timer = Timer.scheduledTimer(timeInterval: timestampRefreshRate, target: TransactionCellWrapper(target: self), selector: NSSelectorFromString("timerDidFire"), userInfo: nil, repeats: true)
-//    } else {
-//        timer?.invalidate()
-//    }
-//    timedateLabel.isHidden = !transaction.isValid
-//
-//    let identity: CGAffineTransform = .identity
-//    if transaction.direction == .received {
-//        arrowImageView.transform = identity.rotated(by: π/2.0)
-//        arrowImageView.tintColor = .txListGreen
-//    } else {
-//        arrowImageView.transform = identity.rotated(by: 3.0*π/2.0)
-//        arrowImageView.tintColor = .cameraGuideNegative
-//    }
-//
-//    if transaction.direction == .received {
-//        qrModalButton.isHidden = false
-//    } else {
-//        qrModalButton.isHidden = true
-//    }
-//}
-//
-
-
-
-//private func configureTransactionCell(transaction:Transaction?, wasSelected: Bool?, indexPath: IndexPath) -> TransactionTableViewCellv2 {
-//
-//    //TODO: Polish animation based on 'wasSelected'
-//
-//    guard let cell = tableView.dequeueReusableCell(withIdentifier: "TransactionTVC2", for: indexPath) as? TransactionTableViewCellv2 else {
-//        NSLog("ERROR: No cell found")
-//        return TransactionTableViewCellv2()
-//    }
-//
-//    if let transaction = transaction {
-//        if transaction.direction == .received {
-//            cell.showQRModalAction = { [unowned self] in
-//
-//                if let addressString = transaction.toAddress,
-//                   let qrImage =  UIImage.qrCode(data: addressString.data(using: .utf8) ?? Data(), color: CIColor(color: .black))?.resize(CGSize(width: qrImageSize, height: qrImageSize)),
-//                   let receiveLTCtoAddressModal = UIStoryboard.init(name: "Alerts", bundle: nil).instantiateViewController(withIdentifier: "LFModalReceiveQRViewController") as? LFModalReceiveQRViewController {
-//
-//                    receiveLTCtoAddressModal.providesPresentationContextTransitionStyle = true
-//                    receiveLTCtoAddressModal.definesPresentationContext = true
-//                    receiveLTCtoAddressModal.modalPresentationStyle = UIModalPresentationStyle.overCurrentContext
-//                    receiveLTCtoAddressModal.modalTransitionStyle = UIModalTransitionStyle.crossDissolve
-//                    receiveLTCtoAddressModal.dismissQRModalAction = { [unowned self] in
-//                        self.dismiss(animated: true, completion: nil)
-//                    }
-//                    self.present(receiveLTCtoAddressModal, animated: true) {
-//                        receiveLTCtoAddressModal.receiveModalTitleLabel.text = S.TransactionDetails.receiveModaltitle
-//                        receiveLTCtoAddressModal.addressLabel.text = addressString
-//                        receiveLTCtoAddressModal.qrImageView.image = qrImage
-//                    }
-//                }
-//            }
-//        }
-//
-//        if let rate = rate,
-//           let store = self.store,
-//           let isLtcSwapped = self.isLtcSwapped {
-//            cell.setTransaction(transaction, isLtcSwapped: isLtcSwapped, rate: rate, maxDigits: store.state.maxDigits, isSyncing: store.state.walletState.syncState != .success)
-//        }
-//
-//        cell.staticBlockLabel.text = S.TransactionDetails.blockHeightLabel
-//        cell.staticCommentLabel.text = S.TransactionDetails.commentsHeader
-//        cell.staticAmountDetailLabel.text = S.Transaction.amountDetailLabel
-//    }
-//    else {
-//        assertionFailure("Transaction must exist")
-//    }
-//    return cell
-//}
-
-
-//    private func configureTransactionCell(transaction:Transaction?, wasSelected: Bool?, indexPath: IndexPath) -> TransactionTableViewCellv2 {
-//
-//        //TODO: Polish animation based on 'wasSelected'
-//
-//        guard let cell = tableView.dequeueReusableCell(withIdentifier: "TransactionTVC2", for: indexPath) as? TransactionTableViewCellv2 else {
-//            NSLog("ERROR: No cell found")
-//            return TransactionTableViewCellv2()
-//        }
-//
-//        if let transaction = transaction {
-//            if transaction.direction == .received {
-//                cell.showQRModalAction = { [unowned self] in
-//
-//                    if let addressString = transaction.toAddress,
-//                        let qrImage =  UIImage.qrCode(data: addressString.data(using: .utf8) ?? Data(), color: CIColor(color: .black))?.resize(CGSize(width: qrImageSize, height: qrImageSize)),
-//                        let receiveLTCtoAddressModal = UIStoryboard.init(name: "Alerts", bundle: nil).instantiateViewController(withIdentifier: "LFModalReceiveQRViewController") as? LFModalReceiveQRViewController {
-//
-//                        receiveLTCtoAddressModal.providesPresentationContextTransitionStyle = true
-//                        receiveLTCtoAddressModal.definesPresentationContext = true
-//                        receiveLTCtoAddressModal.modalPresentationStyle = UIModalPresentationStyle.overCurrentContext
-//                        receiveLTCtoAddressModal.modalTransitionStyle = UIModalTransitionStyle.crossDissolve
-//                        receiveLTCtoAddressModal.dismissQRModalAction = { [unowned self] in
-//                            self.dismiss(animated: true, completion: nil)
-//                        }
-//                        self.present(receiveLTCtoAddressModal, animated: true) {
-//                             receiveLTCtoAddressModal.receiveModalTitleLabel.text = S.TransactionDetails.receiveModaltitle
-//                             receiveLTCtoAddressModal.addressLabel.text = addressString
-//                             receiveLTCtoAddressModal.qrImageView.image = qrImage
-//                         }
-//                    }
-//                }
-//            }
-//
-//            if let rate = rate,
-//                let store = self.store,
-//                let isLtcSwapped = self.isLtcSwapped {
-//                cell.setTransaction(transaction, isLtcSwapped: isLtcSwapped, rate: rate, maxDigits: store.state.maxDigits, isSyncing: store.state.walletState.syncState != .success)
-//            }
-//
-//            cell.staticBlockLabel.text = S.TransactionDetails.blockHeightLabel
-//            cell.staticCommentLabel.text = S.TransactionDetails.commentsHeader
-//            cell.staticAmountDetailLabel.text = S.Transaction.amountDetailLabel
-//        }
-//        else {
-//            assertionFailure("Transaction must exist")
-//        }
-//        return cell
-//    }
-
-
-
-
-//TransactionModalView if let transaction = transaction {
-//        if transaction.direction == .received {
-//            cell.showQRModalAction = { [unowned self] in
-//
-//                if let addressString = transaction.toAddress,
-//                   let qrImage =  UIImage.qrCode(data: addressString.data(using: .utf8) ?? Data(), color: CIColor(color: .black))?.resize(CGSize(width: qrImageSize, height: qrImageSize)),
-//                   let receiveLTCtoAddressModal = UIStoryboard.init(name: "Alerts", bundle: nil).instantiateViewController(withIdentifier: "LFModalReceiveQRViewController") as? LFModalReceiveQRViewController {
-//
-//                    receiveLTCtoAddressModal.providesPresentationContextTransitionStyle = true
-//                    receiveLTCtoAddressModal.definesPresentationContext = true
-//                    receiveLTCtoAddressModal.modalPresentationStyle = UIModalPresentationStyle.overCurrentContext
-//                    receiveLTCtoAddressModal.modalTransitionStyle = UIModalTransitionStyle.crossDissolve
-//                    receiveLTCtoAddressModal.dismissQRModalAction = { [unowned self] in
-//                        self.dismiss(animated: true, completion: nil)
-//                    }
-//                    self.present(receiveLTCtoAddressModal, animated: true) {
-//                        receiveLTCtoAddressModal.receiveModalTitleLabel.text = S.TransactionDetails.receiveModaltitle
-//                        receiveLTCtoAddressModal.addressLabel.text = addressString
-//                        receiveLTCtoAddressModal.qrImageView.image = qrImage
-//                    }
-//                }
-//            }
-//        }
